+++ js-pkg/sdk/src/main.ts
@@ -1,130 +1,11 @@
-/**
- * Schema of object returned after a successful document creation.
- */
-export type DocCreationResult = {
-  /** A unique identifier for the created document. */
-  docId: string
-}
-
-/**
- * An object containing information needed for the client connect to a document.
- *
- * This value is expected to be passed from your server to your client. Your server
- * should obtain this value by calling {@link DocumentManager.getClientToken},
- * and then pass it to the client.
- */
-export type ClientToken = {
-  /** The bare URL of the WebSocket endpoint to connect to. The `doc` string will be appended to this. */
-  url: string
-
-  /** The base URL for document-level endpoints. */
-  baseUrl: string
-
-  /** A unique identifier for the document that the token connects to. */
-  docId: string
-
-  /** A string that grants the bearer access to the document. By default, the development server does not require a token. */
-  token?: string
-}
+import { YSweetError } from './error'
+import type { DocCreationResult, ClientToken, CheckStoreResult } from './types'
+export type { DocCreationResult, ClientToken, CheckStoreResult } from './types'
+export { type YSweetErrorPayload, YSweetError } from './error'
+export { encodeClientToken, decodeClientToken } from './encoding'
 
 function generateRandomString(): string {
   return Math.random().toString(36).substring(2)
-}
-
-/** Metadata associated with a {@link YSweetError}. */
-export type YSweetErrorPayload =
-  | { code: 'ServerRefused'; address: string; port: number; url: string }
-  | { code: 'ServerError'; status: number; message: string; url: string }
-  | { code: 'NoAuthProvided' }
-  | { code: 'InvalidAuthProvided' }
-  | { code: 'Unknown'; message: string }
-
-export type CheckStoreResult = { ok: true } | { ok: false; error: string }
-
-/** An error returned by the y-sweet SDK. */
-export class YSweetError extends Error {
-  /**
-   * Create a new {@link YSweetError}.
-   *
-   * @param cause An object representing metadata associated with the error.
-   * @see {@link YSweetErrorPayload}
-   */
-  constructor(public cause: YSweetErrorPayload) {
-    super(YSweetError.getMessage(cause))
-    this.name = 'YSweetError'
-  }
-
-  /** Convert the message to an error string that can be displayed to the user.
-   *
-   * The error string can also be used with {@link YSweetError.fromMessage} to
-   * reconstruct the payload object, which is useful in the context of Next.js,
-   * which will only pass an error string from the server to the client.
-   *
-   * @param payload The payload object to convert to a string.
-   * @returns A string representation of the error.
-   */
-  static getMessage(payload: YSweetErrorPayload): string {
-    let message
-    if (payload.code === 'ServerRefused') {
-      message = `Server at ${payload.address}:${payload.port} refused connection. URL: ${payload.url}`
-    } else if (payload.code === 'ServerError') {
-      message = `Server responded with ${payload.status} ${payload.message}. URL: ${payload.url}`
-    } else if (payload.code === 'NoAuthProvided') {
-      message = 'No auth provided'
-    } else if (payload.code === 'InvalidAuthProvided') {
-      message = 'Invalid auth provided'
-    } else {
-      message = payload.message
-    }
-    return `${payload.code}: ${message}`
-  }
-
-  /**
-   * In development, next.js passes error objects to the client but strips out everything but the
-   * `message` field. This method allows us to reconstruct the original error object.
-   *
-   * @param messageString The error message string to reconstruct a payload from.
-   * @returns A {@link YSweetError} object.
-   * @see {@link https://nextjs.org/docs/app/api-reference/file-conventions/error#errormessage| Next.js docs}
-   */
-  static fromMessage(messageString: string): YSweetError {
-    let match = messageString.match(/^(.*?): (.*)$/)
-    if (!match) {
-      return new YSweetError({ code: 'Unknown', message: messageString })
-    }
-
-    let [, code, message] = match
-
-    if (code === 'ServerRefused') {
-      match = message.match(/^Server at (.*?):(\d+) refused connection. URL: (.*)$/)
-      if (!match) {
-        return new YSweetError({ code: 'Unknown', message: messageString })
-      }
-
-      let [, address, port, url] = match
-      return new YSweetError({ code, address, port: parseInt(port), url })
-    }
-
-    if (code === 'ServerError') {
-      match = message.match(/^Server responded with (\d+) (.*). URL: (.*)$/)
-      if (!match) {
-        return new YSweetError({ code: 'Unknown', message: messageString })
-      }
-
-      let [, status, statusText, url] = match
-      return new YSweetError({ code, status: parseInt(status), message: statusText, url })
-    }
-
-    if (code === 'NoAuthProvided') {
-      return new YSweetError({ code })
-    }
-
-    if (code === 'InvalidAuthProvided') {
-      return new YSweetError({ code })
-    }
-
-    return new YSweetError({ code: 'Unknown', message })
-  }
 }
 
 /** Represents an interface to a y-sweet document management endpoint. */
@@ -394,44 +275,4 @@ export async function createDoc(
 ): Promise<DocCreationResult> {
   const manager = new DocumentManager(connectionString)
   return await manager.createDoc(docId)
-}
-
-function stringToBase64(input: string) {
-  if (typeof window !== 'undefined' && window.btoa) {
-    // Browser
-    return window.btoa(input)
-  } else if (typeof Buffer !== 'undefined') {
-    // Node.js
-    return Buffer.from(input).toString('base64')
-  } else {
-    throw new Error('Unable to encode to Base64')
-  }
-}
-
-function base64ToString(input: string) {
-  if (typeof window !== 'undefined' && window.atob) {
-    // Browser
-    return window.atob(input)
-  } else if (typeof Buffer !== 'undefined') {
-    // Node.js
-    return Buffer.from(input, 'base64').toString()
-  } else {
-    throw new Error('Unable to decode from Base64')
-  }
-}
-
-export function encodeClientToken(token: ClientToken): string {
-  const jsonString = JSON.stringify(token)
-  let base64 = stringToBase64(jsonString)
-  base64 = base64.replace('+', '-').replace('/', '_').replace(/=+$/, '')
-  return base64
-}
-
-export function decodeClientToken(token: string): ClientToken {
-  let base64 = token.replace('-', '+').replace('_', '/')
-  while (base64.length % 4) {
-    base64 += '='
-  }
-  const jsonString = base64ToString(base64)
-  return JSON.parse(jsonString)
 }
